+++ src/Sentry/Protocol/Envelopes/Envelope.cs
@@ -445,6 +445,22 @@ public sealed class Envelope : ISerializable, IDisposable
         return new Envelope(header, items);
     }
 
+    // TODO: This is temporary. We don't expect single log messages to become an envelope by themselves since batching is needed
+    [Experimental(DiagnosticId.ExperimentalFeature)]
+    internal static Envelope FromLog(SentryLog log)
+    {
+        //TODO: allow batching Sentry logs
+        //see https://github.com/getsentry/sentry-dotnet/issues/4132
+        var header = DefaultHeader;
+
+        var items = new[]
+        {
+            EnvelopeItem.FromLog(log)
+        };
+
+        return new Envelope(header, items);
+    }
+
     private static async Task<IReadOnlyDictionary<string, object?>> DeserializeHeaderAsync(
         Stream stream,
         CancellationToken cancellationToken = default)